+++ changelog.md
@@ -1,1 +1,1 @@
-- 1.21
+- Fixed Mekanism baby enderman not holding blocks correctly